# backend/docs/swagger.yaml
basePath: /api
definitions:
  github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.BadRequestErrorResponse:
    description: Example for a 400 Invalid Input response
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details:
        example: Please check your input and try again
        type: string
      message:
        example: invalid email format
        type: string
    type: object
...
  github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.DetailedErrorResponse:
    description: Detailed error response with code, message and additional details
    properties:
      code:
        allOf:
        - $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.ErrorCode'
        example: VALIDATION_FAILED
      details:
        example: The provided email address is not valid
        type: string
      message:
        example: Invalid email format
        type: string
    type: object
...
  github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.ErrorCode:
    enum:
    - INVALID_REQUEST
    - INVALID_CREDENTIALS
    - VALIDATION_FAILED
    - USER_ALREADY_EXISTS
    - DATABASE_ERROR
    - INTERNAL_ERROR
    - USER_NOT_FOUND
    - TOKEN_INVALID
    - TOKEN_EXPIRED
    - INVALID_SIGNING_METHOD
    - DEVICE_NOT_FOUND
    - DEVICE_ALREADY_EXISTS
    - FORBIDDEN
    type: string
    x-enum-varnames:
    - ErrorCodeInvalidRequest
    - ErrorCodeInvalidCredentials
    - ErrorCodeValidationFailed
    - ErrorCodeUserExists
    - ErrorCodeDatabaseError
    - ErrorCodeInternalError
    - ErrorCodeUserNotFound
    - ErrorCodeTokenInvalid
    - ErrorCodeTokenExpired
    - ErrorCodeInvalidSigningMethod
    - ErrorCodeDeviceNotFound
    - ErrorCodeDeviceAlreadyExists
    - ErrorCodeForbidden
  github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.ForbiddenErrorResponse:
    description: Example for a 403 Invalid Credentials response
    properties:
      code:
        example: FORBIDDEN
        type: string
      details:
        example: Please check your credentials and try again
        type: string
      message:
        example: Invalid credentials
        type: string
    type: object
  github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.HeartbeatResponse:
    description: Heartbeat response with telemetry data and metadata
    properties:
      boot_time:
        example: "2023-01-01T00:00:00Z"
        type: string
      connectivity:
        example: 1
        type: integer
      cpu:
        example: 45.67
        type: number
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      device_id:
        example: f8aac3d7-b2ac-43e8-94d5-bbee59dd4ac9
        type: string
      disk_free:
        example: 23.45
        type: number
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      latency:
        example: 150
        type: integer
      ram:
        example: 67.89
        type: number
      temperature:
        example: 35.67
        type: number
    type: object
  github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.InternalServerErrorResponse:
    description: Example for a 500 Internal Server Error response
    properties:
      code:
        example: INTERNAL_SERVER_ERROR
        type: string
      details:
        example: failed to create user
        type: string
      message:
        example: Internal server error
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for DT Labs Fullstack Exam - Device Management System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: DT Labs Exam API
  version: "1.0"
paths:
...
  /v1/devices/{id}/heartbeats:
    get:
      consumes:
      - application/json
      description: Get heartbeats for a specific device within a time range
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - default: 24 hours ago
        description: Start time (RFC3339 format)
        in: query
        name: start
        type: string
      - default: now
        description: End time (RFC3339 format)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of device heartbeats
          schema:
            items:
              $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.HeartbeatResponse'
            type: array
        "400":
          description: Invalid device ID or time format
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.BadRequestErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.DetailedErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.ForbiddenErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.DetailedErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get device heartbeats
      tags:
      - heartbeats
  /v1/devices/{id}/heartbeats/latest:
    get:
      consumes:
      - application/json
      description: Get the most recent heartbeat for a specific device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latest device heartbeat
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.HeartbeatResponse'
        "400":
          description: Invalid device ID
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.BadRequestErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.DetailedErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.ForbiddenErrorResponse'
        "404":
          description: Device or heartbeat not found
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.DetailedErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Arthur-7Melo_exame-fullstack-setembro-dtlabs-2025_internal_dto.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get latest device heartbeat
      tags:
      - heartbeats
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"